Alert Condition phase1

+++ Delfi.Glo.Entities/Dto/WellDto.cs
@@ -27,10 +27,10 @@ namespace Delfi.Glo.Entities.Dto
         public string ApprovalStatus { get; set; }    
         public string UserId { get; set; }
         public int NoOfAlerts { get; set; }
-        public decimal GasLiftInjectionRate { get; set; }
-        public decimal DeltaPressure { get; set; }
-        public decimal TubingHeadPressure { get; set; }
-        public decimal FlowLinePressure { get; set; }
-        public decimal CasingHeadPressure { get; set; }
+        public int GasLiftInjectionRate { get; set; }
+        public int DeltaPressure { get; set; }
+        public int TubingHeadPressure { get; set; }
+        public int FlowLinePressure { get; set; }
+        public int CasingHeadPressure { get; set; }
     }
 }
